finished chessboard and all piece classes, fixed coordinates, fixed every move

+++ src/Chessboard.java
@@ -5,55 +5,51 @@ public class Chessboard {
 
     Chessboard() {
         board[0][0] = new Rook(0, 0, false);
-        board[0][1] = new Bishop(0, 1, false);
-        board[0][2] = new Knight(0, 2, false);
-        board[0][3] = new Queen(0, 3, false);
-        board[0][4] = new King(0, 4, false);
-        board[0][5] = new Knight(0, 5, false);
-        board[0][6] = new Bishop(0, 6, false);
-        board[0][7] = new Rook(0, 7, false);
+        board[1][0] = new Bishop(1, 0, false);
+        board[2][0] = new Knight(2, 0, false);
+        board[3][0] = new Queen(3, 0, false);
+        board[4][0] = new King(4, 0, false);
+        board[5][0] = new Knight(5, 0, false);
+        board[6][0] = new Bishop(6, 0, false);
+        board[7][0] = new Rook(7, 0, false);
         for (int i = 0; i < 8; i++) {
-            board[1][i] = new Pawn(1, i, false);
+            board[i][1] = new Pawn(i, 1, false);
         }
 
-        board[7][0] = new Rook(0, 0, true);
-        board[7][1] = new Bishop(0, 1, true);
-        board[7][2] = new Knight(0, 2, true);
-        board[7][3] = new Queen(0, 3, true);
-        board[7][4] = new King(0, 4, true);
-        board[7][5] = new Knight(0, 5, true);
-        board[7][6] = new Bishop(0, 6, true);
-        board[7][7] = new Rook(0, 7, true);
+        board[0][7] = new Rook(0, 7, true);
+        board[1][7] = new Bishop(1, 7, true);
+        board[2][7] = new Knight(2, 7, true);
+        board[3][7] = new Queen(3, 7, true);
+        board[4][7] = new King(4, 7, true);
+        board[5][7] = new Knight(5, 7, true);
+        board[6][7] = new Bishop(6, 7, true);
+        board[7][7] = new Rook(7, 7, true);
         for (int i = 0; i < 8; i++) {
-            board[6][i] = new Pawn(1, i, true);
+            board[i][6] = new Pawn(i, 6, true);
         }
     }
 
-    public void move(int currentX, int currentY, int newX, int newY) {
+    public void move(int currentX, int currentY, int newX, int newY) throws IllegalArgumentException, IllegalStateException {
         Piece currentPiece = board[currentX][currentY];
+        if (currentPiece == null) {
+            return;
+        }
         if (board[newX][newY] == null) {
-            try {
-                currentPiece.move(newX, newY);
-                board[currentX][currentY] = null;
-                board[newX][newY] = currentPiece;
-            } catch (IllegalArgumentException e) {
-                System.out.println(e.getMessage());
-            }
+            currentPiece.move(newX, newY, board);
+            board[currentX][currentY] = null;
+            board[newX][newY] = currentPiece;
+
         } else {
             if (currentPiece.isPlayerOne() != board[newX][newY].isPlayerOne()) {
-                try {
-                    if (currentPiece instanceof Pawn) {
-                        ((Pawn) currentPiece).kill(newX, newY);
-                    } else {
-                        currentPiece.move(newX, newY);
-                    }
-                    board[currentX][currentY] = null;
-                    board[newX][newY] = currentPiece;
-                } catch (IllegalArgumentException e) {
-                    System.out.println(e.getMessage());
+                if (currentPiece instanceof Pawn) {
+                    ((Pawn) currentPiece).kill(newX, newY);
+                } else {
+                    currentPiece.move(newX, newY, board);
                 }
+                board[currentX][currentY] = null;
+                board[newX][newY] = currentPiece;
             } else {
-                throw new IllegalArgumentException();
+                throw new IllegalArgumentException("Cant kill own Pawn");
             }
         }
     }
@@ -63,7 +59,7 @@ public class Chessboard {
         String out = "";
         for (int i = 0; i < 8; i++) {
             for (int j = 0; j < 8; j++) {
-                out += String.format("%8s \t", board[i][j] == null ? "--------" : board[i][j].toString());
+                out += String.format("%8s \t", board[j][i] == null ? "--------" : board[j][i].toString());
             }
             out += "\n";
         }
